@@ -8,19 +8,19 @@ use CSSFromHTMLExtractor\Css\Rule\Rule;
 class CssStore
 {
     /** @var array Property objects, grouped by selector */
-    private $styles = [];
+    private array $styles = [];
 
     /** @var string|null */
-    private $charset;
+    private ?string $charset;
 
-    public function addCssStyles($cssRules)
+    public function addCssStyles($cssRules): static
     {
         $this->styles = array_merge($this->styles, $cssRules);
 
         return $this;
     }
 
-    public function getStyles()
+    public function getStyles(): array
     {
         return $this->styles;
     }
@@ -28,7 +28,7 @@ class CssStore
     /**
      * @return $this
      */
-    public function purge()
+    public function purge(): static
     {
         $this->styles = [];
 
@@ -40,12 +40,12 @@ class CssStore
      *
      * @return bool whether the dumping was successful
      */
-    public function dumpStyles($path)
+    public function dumpStyles($path): bool
     {
         return file_put_contents($path, $this->compileStyles()) === false;
     }
 
-    public function compileStyles()
+    public function compileStyles(): string
     {
         // Structure rules in order, by media query
         $styles = $this->prepareStylesForProcessing();
@@ -73,7 +73,7 @@ class CssStore
      * @return string
      *
      */
-    private function parseMediaToString($media, array $rules)
+    private function parseMediaToString($media, array $rules): string
     {
         if ($media == '') {
             return
@@ -107,7 +107,7 @@ class CssStore
      *
      * @return string
      */
-    private function parsePropertiesToString($selector, array $properties)
+    private function parsePropertiesToString($selector, array $properties): string
     {
         return "$selector { " .
             join(
@@ -122,7 +122,7 @@ class CssStore
             "}";
     }
 
-    private function prepareStylesForProcessing()
+    private function prepareStylesForProcessing(): array
     {
         // Group styles by order and media
         $groupedStyles = [];
@@ -135,7 +135,7 @@ class CssStore
         return $groupedStyles;
     }
 
-    public function setCharset($charset)
+    public function setCharset($charset): void
     {
         $this->charset = $charset;
     }
